pyproject.toml
[tool.poetry]
name = "calcengine"
version = "0.1.0"
description = "A Lazy Graph Calculation Engine"
authors = ["Blair Azzopardi <blairuk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pyqt5 = { version = "^5.15.0", optional = true }
pandas = { version = "^1.0.5", optional = true }
matplotlib = { version = "^3.3.0", optional = true }
Pillow = { version = "^7.2.0", optional = true }

[tool.poetry.extras]
demo = ["pyqt5", "pandas", "matplotlib", "Pillow"]

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.3"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
